+++ cpp/src/unity/kh_plugin.cpp
@@ -25,7 +25,7 @@ extern "C"
         int frame_size
     )
     {
-        return new kh::FFmpegFrame(std::move(decoder->decode({frame_data, frame_size})));
+        return new kh::FFmpegFrame(std::move(decoder->decode({frame_data, gsl::narrow_cast<size_t>(frame_size)})));
     }
 
     UNITY_INTERFACE_EXPORT void UNITY_INTERFACE_API delete_ffmpeg_frame(kh::FFmpegFrame* ptr)
@@ -51,7 +51,7 @@ extern "C"
         bool keyframe
     )
     {
-        return new std::vector<std::int16_t>(std::move(decoder->decode({frame_data, frame_size}, keyframe)));
+        return new std::vector<std::int16_t>(std::move(decoder->decode({frame_data, gsl::narrow_cast<size_t>(frame_size)}, keyframe)));
     }
 
     UNITY_INTERFACE_EXPORT void UNITY_INTERFACE_API delete_depth_pixels(std::vector<int16_t>* ptr)
@@ -79,7 +79,7 @@ extern "C"
     )
     {
         if(opus_frame_data)
-            return decoder->decode(gsl::span<std::byte>{opus_frame_data, opus_frame_size}, pcm_data, frame_size, 0);
+            return decoder->decode(gsl::span<std::byte>{opus_frame_data, gsl::narrow_cast<size_t>(opus_frame_size)}, pcm_data, frame_size, 0);
         else
             return decoder->decode(std::nullopt, pcm_data, frame_size, 0);
     }
